Add files via upload
This paper describes a statistical model used to forecast the 2016 presidential election at the county level. We combined census data with election data for each county in the US, and conducted principal component analysis on a training set. We built a random forest model atop the principal components, which produced accurate forecasts of the test set.